refactor(recipe): :card_file_box: change recipe prep time type to int

+++ src/recipe/entities/recipe.entity.ts
@@ -42,7 +42,7 @@ export class Recipe {
   @Column({ type: 'text' })
   description: string;
 
-  @Column({ type: 'double' })
+  @Column({ type: 'int' })
   preparationTimeInMinutes: number;
 
   @Column({ type: 'enum', enum: RecipeSize, default: RecipeSize.Small })
